This commit includes necessary breaking changes.
The DSL as originally designed made it necessary to use variables in
many cases to refer to elements in views when the elements were not in
the view scope. The DSL also could create subtle issues where names
would not resolve to the intended element in some cases (again forcing
the use of variables).
This change introduces a few changes to make scoping unambiguous.
The commit also fixes potential race conditions in how elements were
added to or removed from views. Generation is now done in two passes to
guarantees elements are added and removed in the appropriate order
independently from the order they appear in the DSL.
The commit also includes changes to the interface between Structurizr
and the output of the DSL which results in a cleaner model.
Structurizr specific constructs and business logic is now entirely
encapsulated in the `stz` package.

List of breaking changes in the DSL:

* Workspace -> Design
* Remove is now replaced with Remove, RemoveTagged and Unlink.
* DSL functions such as Add, Link, Remove and Unlink that accept
  element names now also accept element paths. The path of an
  element is built by concatenating the top level scopes with the
  element name separated with slashes
  (e.g. "<Software System/<Container>/<Component>")
* Views now require a key.