Remove External tag and fix potential crashes (#706)
The External tag does not allow styling because of its position. Also
different external systems may need to be styled differently so revert
the External tag creation and instead rely on explicit tags in the design.

Fix a few missing nil checks both in the JS app and in Go.